\name{r4test-package}
\alias{r4test-package}
\alias{r4test}
\docType{package}
\title{
\packageTitle{r4test}
}
\description{
\packageDescription{r4test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{r4test}
\packageIndices{r4test}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{r4test}

Maintainer: \packageMaintainer{r4test}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
